Rename move_or copy modal file to reflect the new Select action
